requirements.txt
accelerate==1.2.1
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
executing==2.1.0
filelock==3.16.1
-e git+https://github.com/foundation-model-stack/fms-extras@16339f7c82255983d20dabf4807d124133be1d1c#egg=fms_extras
fonttools==4.55.3
fsspec==2024.10.0
-e git+https://github.com/kushaangowda/nestedtensors-for-transformers.git@23958219c9bd40cf77de9c0b7497fdf8380f5d4c#egg=hpml_utils&subdirectory=hpml_utils
huggingface-hub==0.26.5
ibm-fms==0.0.8
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pillow==11.0.0
print-torch==1.3
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.17.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.47.0
triton==2.2.0
typing_extensions==4.12.2
urllib3==2.2.3
varname==0.13.5
